indexer-agent,-common: Use consistent clear network/epoch var naming
