+++ packages/indexer-common/src/indexer-management/index.ts
@@ -6,3 +6,4 @@ export * from './monitor'
 export * from './server'
 export * from './subgraphs'
 export * from './rules'
+export * from './types'
